+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/Amqp/AmqpMessageConverter.cs
@@ -20,11 +20,6 @@ namespace Azure.Messaging.ServiceBus.Amqp
 {
     internal class AmqpMessageConverter
     {
-        /// <summary>
-        /// The size, in bytes, to use for extracting the delivery tag bytes into <see cref="Guid"/>.
-        /// </summary>
-        private const int GuidSizeInBytes = 16;
-
         /// <summary>The size, in bytes, to use as a buffer for stream operations.</summary>
         private const int StreamBufferSizeInBytes = 512;
 
@@ -558,37 +553,11 @@ namespace Azure.Messaging.ServiceBus.Amqp
 
             // lock token
 
-            sbMessage.LockTokenGuid = ParseGuidBytes(amqpMessage.DeliveryTag);
+            sbMessage.LockTokenGuid = GuidUtilities.ParseGuidBytes(amqpMessage.DeliveryTag);
 
             amqpMessage.Dispose();
 
             return sbMessage;
-        }
-
-        public static Guid ParseGuidBytes(ReadOnlyMemory<byte> bytes)
-        {
-            if (bytes.Length == GuidSizeInBytes)
-            {
-                // Use TryRead to avoid allocating an array if we are on a little endian machine.
-                if (!BitConverter.IsLittleEndian || !MemoryMarshal.TryRead<Guid>(bytes.Span, out var lockTokenGuid))
-                {
-                    // Either we are on a big endian machine or the bytes were not a valid GUID.
-                    // Even if the bytes were not valid, use the Guid constructor to leverage the Guid validation rather than throwing ourselves.
-                    lockTokenGuid = new Guid(bytes.ToArray());
-                }
-                return lockTokenGuid;
-            }
-
-            return default;
-        }
-
-        public static void WriteGuidBytes(Guid guid, byte[] buffer)
-        {
-            Debug.Assert(buffer is { Length: 16 });
-            if (!BitConverter.IsLittleEndian || !MemoryMarshal.TryWrite(buffer, ref guid))
-            {
-                guid.ToByteArray().AsSpan().CopyTo(buffer);
-            }
         }
 
         internal static bool TryGetAmqpObjectFromNetObject(object netObject, MappingType mappingType, out object amqpObject)
